- Stencil clipping improvemens - IMGUI rendering improvements

+++ rawrbox.render/src/renderer.cpp
@@ -277,7 +277,7 @@ namespace rawrbox {
 		if (rawrbox::CHECKER_TEXTURE == nullptr) {
 			auto checker = rawrbox::TextureUtils::generateCheckboard({256U, 256U}, rawrbox::Color::RGBHex(0x161618), rawrbox::Color::RGBHex(0x2a2a2d), 8U);
 
-			rawrbox::CHECKER_TEXTURE = std::make_shared<rawrbox::TextureImage>(rawrbox::Vector2u(256U, 256U), checker, 4);
+			rawrbox::CHECKER_TEXTURE = std::make_shared<rawrbox::TextureImage>(rawrbox::Vector2u(256U, 256U), checker, uint8_t(4));
 			rawrbox::CHECKER_TEXTURE->setName("CHECKER_TEXTURE");
 			rawrbox::CHECKER_TEXTURE->upload();
 		}
@@ -288,15 +288,15 @@ namespace rawrbox {
 		this->_logger->info("Loading default fonts");
 
 		if (rawrbox::DEBUG_FONT_REGULAR == nullptr) {
-			rawrbox::DEBUG_FONT_REGULAR = rawrbox::TextEngine::load("./assets/fonts/SourceCodePro-Regular.ttf", 11);
+			rawrbox::DEBUG_FONT_REGULAR = rawrbox::TextEngine::load("./assets/fonts/SourceCodePro-Regular.ttf", 11U);
 		}
 
 		if (rawrbox::DEBUG_FONT_BOLD == nullptr) {
-			rawrbox::DEBUG_FONT_BOLD = rawrbox::TextEngine::load("./assets/fonts/SourceCodePro-Bold.ttf", 11);
+			rawrbox::DEBUG_FONT_BOLD = rawrbox::TextEngine::load("./assets/fonts/SourceCodePro-Bold.ttf", 11U);
 		}
 
 		if (rawrbox::DEBUG_FONT_ITALIC == nullptr) {
-			rawrbox::DEBUG_FONT_ITALIC = rawrbox::TextEngine::load("./assets/fonts/SourceCodePro-Italic.ttf", 11);
+			rawrbox::DEBUG_FONT_ITALIC = rawrbox::TextEngine::load("./assets/fonts/SourceCodePro-Italic.ttf", 11U);
 		}
 		// ------
